Add aggregated progress
In the FileProcessor we now process the progress of the individual
downloads and create an aggregated progress, which is then reported back
to main (instead of the detailed progress).
The aggregated progress keeps the last reported FileProgress instance as
well as an overall progress indicator.
The FileProcessor keeps the percentage complete for each file, in order
to be able to calculate the overall progress.