Make process run on threadpool threads
Since we did not do a Task.Run() to execute the "download" tasks, the
tasks would simply run on the UI thread. This worked fine, but was not
what we wanted.
Put in the Task.Run() statement, and added extra information to the
progress class, so that we can display the thread id in the UI and
verify that the tasks are in fact running on threadpool threads.